@@ -41,8 +41,12 @@
     "@types/sodium-native": "^2.3.5",
     "@types/workerpool": "^6.0.0",
     "@types/yn": "^3.1.0",
+    "@typescript-eslint/eslint-plugin": "^4.31.2",
+    "@typescript-eslint/parser": "^4.31.2",
+    "eslint": "^7.32.0",
     "jest": "^27.0.6",
+    "prettier": "^2.4.1",
     "ts-jest": "^27.0.4",
-    "typescript": "^4.3.4"
+    "typescript": "^4.4.3"
   }
 }
