@@ -1,23 +1,20 @@
 import {
     createSCPEnvelopeSignature,
-    createStatementXDRSignature,
     verifySCPEnvelopeSignature
-} from "../src/stellar-message-service";
-import {xdr, Keypair, Networks, hash, sign} from "stellar-base";
+} from "../src";
+import {xdr, Keypair, Networks, hash} from "stellar-base";
 
 test("It should create and verify envelope signatures correctly", ()=>{
-    let keyPair = Keypair.random();
-    console.log(keyPair.secret());
-    console.log(keyPair.rawSecretKey());
-    let commit = new xdr.ScpBallot({counter: 1, value: Buffer.alloc(32)});
-    let externalize = new xdr.ScpStatementExternalize({
+    const keyPair = Keypair.random();
+    const commit = new xdr.ScpBallot({counter: 1, value: Buffer.alloc(32)});
+    const externalize = new xdr.ScpStatementExternalize({
         commit: commit,
         nH: 1,
         commitQuorumSetHash: Buffer.alloc(32)
     })
-    let pledges = xdr.ScpStatementPledges.scpStExternalize(externalize);
+    const pledges = xdr.ScpStatementPledges.scpStExternalize(externalize);
 
-    let statement = new xdr.ScpStatement(
+    const statement = new xdr.ScpStatement(
         {
             nodeId: xdr.PublicKey.publicKeyTypeEd25519(keyPair.rawPublicKey()),
             slotIndex: xdr.Uint64.fromString("1"),
@@ -25,12 +22,12 @@ test("It should create and verify envelope signatures correctly", ()=>{
         }
     )
 
-    let signatureResult = createSCPEnvelopeSignature(statement, keyPair.rawPublicKey(), keyPair.rawSecretKey(), hash(Buffer.from(Networks.PUBLIC)));
+    const signatureResult = createSCPEnvelopeSignature(statement, keyPair.rawPublicKey(), keyPair.rawSecretKey(), hash(Buffer.from(Networks.PUBLIC)));
     if(signatureResult.isErr()){
         expect(signatureResult.isErr()).toBeFalsy();
     } else {
-        let envelope = new xdr.ScpEnvelope({statement: statement, signature: signatureResult.value})
-        let result = verifySCPEnvelopeSignature(envelope, hash(Buffer.from(Networks.PUBLIC)));
+        const envelope = new xdr.ScpEnvelope({statement: statement, signature: signatureResult.value})
+        const result = verifySCPEnvelopeSignature(envelope, hash(Buffer.from(Networks.PUBLIC)));
         expect(result.isOk()).toBeTruthy();
         if(result.isOk())
             expect(result.value).toBeTruthy();
